requirements.txt
streamlit>=1.33.0
opencv-python>=4.9.0
numpy>=1.26.0
pillow>=10.0.0
torch>=2.2.0
torchaudio>=2.2.0
torchvision>=0.17.0
fastsam>=0.1.0
